refactor(sys): extract Windows symbol module to separate file
